+++ pypdf/_image_xobject.py
